backend/requirements.txt
Django==4.2.13
djangorestframework==3.15.2
psycopg2-binary==2.9.9
kafka-python==2.0.2
pandas==2.2.2
numpy==2.0.1
gunicorn==22.0.0